(1)新增商品时，发布商品，并将商品指定信息存入索引库。(ActiveMQ)

+++ dd-manager/dd-manager-interface/src/main/java/com/qjk/ddshop/service/ItemService.java
@@ -1,5 +1,6 @@
 package com.qjk.ddshop.service;
 
+import com.qjk.ddshop.common.dto.MessageResult;
 import com.qjk.ddshop.common.dto.Order;
 import com.qjk.ddshop.common.dto.Page;
 import com.qjk.ddshop.common.dto.Result;
@@ -25,5 +26,5 @@ public interface ItemService {
      * @param paramData JSON格式字符串，存商品规格数据
      * @return  新增商品条数
      */
-    int saveItem(TbItem tbItem, String content,String paramData);
+    Long saveItem(TbItem tbItem, String content, String paramData);
 }
